fix: rename blog-post template into pattern-page, handle no tabs state, and rename subtitle presentation if index page
