+++ src/gatsby-components/Hint/Hint.js
