+++ src/gatsby-components/Hint/index.js
